update get all users handler and service
